+++ src/CheckIn.js
@@ -326,6 +326,7 @@ class CheckIn extends React.Component {
       },
       'barcode': loan => `${get(loan, ['item', 'barcode'])}`,
       'location': loan => `${get(loan, ['item', 'location', 'name'])}`,
+      'inHouseUse': loan => { return get(loan, 'inHouseUse') ? <Icon icon="plus-sign" size="small" /> : ''; },
       'status': loan => {
         const status = `${get(loan, ['item', 'status', 'name'])}`;
         const inTransitSp = get(loan, ['item', 'inTransitDestinationServicePoint', 'name']);
@@ -345,6 +346,7 @@ class CheckIn extends React.Component {
       'barcode': formatMessage({ id: 'ui-checkin.barcode' }),
       'effectiveCallNumber': formatMessage({ id: 'ui-checkin.effectiveCallNumber' }),
       'location': formatMessage({ id: 'ui-checkin.location' }),
+      'inHouseUse': formatMessage({ id: 'ui-checkin.inHouseUse' }),
       'status': formatMessage({ id: 'ui-checkin.status' }),
       ' ': <IconButton style={{ marginLeft: '-6px' }} icon="gear" aria-label="action settings" />,
     };
@@ -459,9 +461,9 @@ class CheckIn extends React.Component {
                   <MultiColumnList
                     id="list-items-checked-in"
                     fullWidth
-                    visibleColumns={['timeReturned', 'title', 'barcode', 'effectiveCallNumber', 'location', 'status', ' ']}
+                    visibleColumns={['timeReturned', 'title', 'barcode', 'effectiveCallNumber', 'location', 'inHouseUse', 'status', ' ']}
                     columnMapping={columnMapping}
-                    columnWidths={{ 'timeReturned': 120, ' ': 80, 'title': 300, 'barcode': 200, 'effectiveCallNumber': 200, 'location': 200, 'status': 120 }}
+                    columnWidths={{ 'timeReturned': 120, ' ': 80, 'title': 300, 'barcode': 200, 'effectiveCallNumber': 200, 'location': 200, 'inHouseUse': 80, 'status': 120 }}
                     columnOverflow={{ ' ': true }}
                     rowMetadata={['id']}
                     interactive={false}