+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Display effective call number prefix, call number, and call number suffix at check in. Refs UICHKIN-127.
 * Update okapiInterfaces: `item-storage: 8.0`, `circulation: 9.0`. Refs UICHKIN-149.
 * Make the Check in ellipsis accessible. Refs UICHKIN-134.
+* Add 'in-house use' column to checked-in items table. Refs UICHKIN-100.
 
 ## [1.10.0](https://github.com/folio-org/ui-checkin/tree/v1.10.0) (2019-12-4)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.9.0...v1.10.0)