+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Make the Check in ellipsis accessible. Refs UICHKIN-134.
 * Add 'in-house use' column to checked-in items table. Refs UICHKIN-100.
 * Security update eslint to >= 6.2.1 or eslint-util >= 1.4.1. Refs. UICHKIN-150.
+* Add declared lost modal. Refs UICHKIN-114.
 
 ## [1.10.0](https://github.com/folio-org/ui-checkin/tree/v1.10.0) (2019-12-4)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v1.9.0...v1.10.0)